remove old scanpy-like function
